fishing_spot v0.0.1 FishingSpot.QueryExamples.Selects

These examples demonstrate the different styles of Ecto select clauses.

Functions

big_query()

iex> FishingSpot.QueryExamples.Selects.big_query

    from fish in FishLanded,
      join: fly_type in assoc(fish, :fly_type),
      join: fish_species in assoc(fish, :fish_species),
      join: fisherman in assoc(fish, :fisherman),
      join: trip in assoc(fisherman, :trips),
      join: locations in assoc(trip, :locations),
      join: location_types in assoc(locations, :location_type),
      where: fragment(
        "? IN (SELECT MAX(biggest_fish.length) FROM fish_landed biggest_fish)", fish.length
      ),
      where: fish.date_and_time >= trip.start_date,
      where: fish.date_and_time <= date_add(trip.end_date, 1, "day"),
      select: %{
        length: fish.length,
        date_caught: fish.date_and_time,
        fish_type: fish_species.name,
        fly: fly_type.name,
        fisherman: fisherman.name,
        trip_start: trip.start_date,
        trip_end: trip.end_date,
        location: locations.name,
        location_type: location_types.name
      }

SELECT f0."length", f0."date_and_time", f2."name", f1."name", f3."name", t4."start_date", t4."end_date", l5."name", l6."name" FROM "fish_landed" AS f0 INNER JOIN "fly_types" AS f1 ON f1."id" = f0."fly_type_id" INNER JOIN "fish_species" AS f2 ON f2."id" = f0."fish_species_id" INNER JOIN "fishermen" AS f3 ON f3."id" = f0."fisherman_id" INNER JOIN "fishermen_trips" AS f7 ON f7."fisherman_id" = f3."id" INNER JOIN "trips" AS t4 ON t4."id" = f7."trip_id" INNER JOIN "locations_trips" AS l8 ON l8."trip_id" = t4."id" INNER JOIN "locations" AS l5 ON l5."id" = l8."location_id" INNER JOIN "location_types" AS l6 ON l6."id" = l5."location_type_id" WHERE (f0."length" IN (SELECT MAX(biggest_fish.length) FROM fish_landed biggest_fish)) AND (f0."date_and_time" >= t4."start_date") AND (f0."date_and_time" <= (t4."end_date"::date + interval '1 day')::date)

[]

default_return()

By default an Ecto query will return a List of structs. The struct will be the type of the schema given in the from/2 clause ( Ecto.Query.from/2 ). In this case they will be %FishingSpot.Fisherman{} structs.

iex> FishingSpot.QueryExamples.Selects.default_return

    from fisherman in Fisherman

SELECT f0."id", f0."inserted_at", f0."updated_at", f0."name", f0."date_of_birth" FROM "fishermen" AS f0

[%FishingSpot.Fisherman{__meta__: #Ecto.Schema.Metadata<:loaded, "fishermen">,
  date_of_birth: ~D[1970-01-01],
  fish_landed: #Ecto.Association.NotLoaded<association :fish_landed is not loaded>,
  fishermen_trips: #Ecto.Association.NotLoaded<association :fishermen_trips is not loaded>,
  ...},
 %FishingSpot.Fisherman{__meta__: #Ecto.Schema.Metadata<:loaded, "fishermen">,
  date_of_birth: ~D[1978-03-05],
  fish_landed: #Ecto.Association.NotLoaded<association :fish_landed is not loaded>,
  ...},
 %FishingSpot.Fisherman{__meta__: #Ecto.Schema.Metadata<:loaded, "fishermen">,
  date_of_birth: ~D[1973-10-15], ...},
 %FishingSpot.Fisherman{__meta__: #Ecto.Schema.Metadata<:loaded, "fishermen">,
  ...}, %FishingSpot.Fisherman{...}, ...]